Create MaximumNumber.cpp
Program which returns an array of the k digits of maximum number of length k <= m + n from digits of the two numbers. , given two integer arrays nums1 and nums2 of lengths m and n respectively. nums1 and nums2 represent the digits of two numbers. Also given an integer k.